// src/lib/types/riot.type.ts
export const RuneGroup = z.union([
    z.literal(0), // none
    z.literal(8000), // domination
    z.literal(8100), // precision
    z.literal(8200), // sorcery
    z.literal(8300), // whimsy
    z.literal(8400), // resolve
])

/**
 * Rune Perk - the runes available to select
 */
export const RunePerk = z.union([
    z.literal(0), // none
    z.literal(8005),
    z.literal(8008),
    z.literal(8009),
    z.literal(8010),
    z.literal(8014),
    z.literal(8017),
    z.literal(8021),
    z.literal(8105),
    z.literal(8106),
    z.literal(8112),
    z.literal(8120),
    z.literal(8124),
    z.literal(8128),
    z.literal(8126),
    z.literal(8134),
    z.literal(8135),
    z.literal(8136),
    z.literal(8138),
    z.literal(8139),
    z.literal(8143),
    z.literal(8210),
    z.literal(8214),
    z.literal(8224),
    z.literal(8226),
    z.literal(8229),
    z.literal(8230),
    z.literal(8232),
    z.literal(8233),
    z.literal(8234),
    z.literal(8236),
    z.literal(8237),
    z.literal(8242),
    z.literal(8275),
    z.literal(8299),
    z.literal(8304),
    z.literal(8306),
    z.literal(8313),
    z.literal(8316),
    z.literal(8321),
    z.literal(8345),
    z.literal(8347),
    z.literal(8351),
    z.literal(8352),
    z.literal(8360),
    z.literal(8369),
    z.literal(8401),
    z.literal(8410),
    z.literal(8429),
    z.literal(8437),
    z.literal(8439),
    z.literal(8444),
    z.literal(8446),
    z.literal(8451),
    z.literal(8453),
    z.literal(8463),
    z.literal(8465),
    z.literal(8473),
    z.literal(9101),
    z.literal(9103),
    z.literal(9104),
    z.literal(9105),
    z.literal(9111),
    z.literal(9923),
])
// ...
const RiotObjectiveSchema = z.object({
    first: z.boolean(),
    kills: z.number(),
})

const RiotTeamSchema = z.object({
    bans: z.array(
        z.object({
            championId: z.number(),
            pickTurn: z.number(),
        }),
    ),
    objectives: z.object({
        baron: RiotObjectiveSchema,
        champion: RiotObjectiveSchema,
        dragon: RiotObjectiveSchema,
        inhibitor: RiotObjectiveSchema,
        riftHerald: RiotObjectiveSchema,
        tower: RiotObjectiveSchema,
    }),
    teamId: z.union([z.literal(100), z.literal(200), z.literal(0)]),
    win: z.boolean(),
})

export const RiotGameSchema = z.object({
    metadata: z.object({
        dataVersion: z.string(),
        matchId: z.string(),
        participants: z.array(z.string()),
    }),
    info: z.object({
        gameCreation: z.number(),
        gameDuration: z.number(),
        gameEndTimestamp: z.number(),
        gameId: z.number(),
        gameMode: z.string(),
        gameName: z.string(),
        gameStartTimestamp: z.number(),
        gameType: z.string(),
        gameVersion: z.string(),
        mapId: z.number(),
        participants: z.array(RiotParticipantSchema),
        platformId: z.string(),
        queueId: z.number(),
        teams: z.tuple([RiotTeamSchema, RiotTeamSchema]),
    }),
})

export type RiotGameType = z.infer<typeof RiotGameSchema>